[SPARK-47377][PYTHON][CONNECT][TESTS] Factor out tests from `SparkConnectSQLTestCase`
### What changes were proposed in this pull request?
Factor out tests from `SparkConnectSQLTestCase`

### Why are the changes needed?
for testing parallelism

### Does this PR introduce _any_ user-facing change?
no, test only

### How was this patch tested?
ci

### Was this patch authored or co-authored using generative AI tooling?
no

Closes #45497 from zhengruifeng/breaK_spark_connect_basic.

Authored-by: Ruifeng Zheng <ruifengz@apache.org>
