+++ python/pyspark/sql/tests/connect/test_connect_basic.py
@@ -15,13 +15,9 @@
 # limitations under the License.
 #
 
-import array
-import datetime
 import os
 import unittest
-import random
 import shutil
-import string
 import tempfile
 
 from pyspark.errors import (
@@ -43,7 +39,6 @@ from pyspark.sql.types import (
 )
 
 from pyspark.testing.sqlutils import (
-    MyObject,
     SQLTestUtils,
     PythonOnlyUDT,
     ExamplePoint,
@@ -56,17 +51,13 @@ from pyspark.testing.connectutils import (
 from pyspark.testing.pandasutils import PandasOnSparkTestUtils
 from pyspark.errors.exceptions.connect import (
     AnalysisException,
-    ParseException,
     SparkConnectException,
 )
 
 if should_test_connect:
-    import pandas as pd
-    import numpy as np
     from pyspark.sql.connect.proto import Expression as ProtoExpression
     from pyspark.sql.connect.session import SparkSession as RemoteSparkSession
     from pyspark.sql.connect.column import Column
-    from pyspark.sql.connect.readwriter import DataFrameWriterV2
     from pyspark.sql.dataframe import DataFrame
     from pyspark.sql.connect.dataframe import DataFrame as CDataFrame
     from pyspark.sql import functions as SF
@@ -274,165 +265,6 @@ class SparkConnectBasicTests(SparkConnectSQLTestCase):
         with self.assertRaises(AnalysisException):
             df.collect()
 
-    def test_simple_read(self):
-        df = self.connect.read.table(self.tbl_name)
-        data = df.limit(10).toPandas()
-        # Check that the limit is applied
-        self.assertEqual(len(data.index), 10)
-
-    def test_json(self):
-        with tempfile.TemporaryDirectory(prefix="test_json") as d:
-            # Write a DataFrame into a JSON file
-            self.spark.createDataFrame([{"age": 100, "name": "Hyukjin Kwon"}]).write.mode(
-                "overwrite"
-            ).format("json").save(d)
-            # Read the JSON file as a DataFrame.
-            self.assert_eq(self.connect.read.json(d).toPandas(), self.spark.read.json(d).toPandas())
-
-            for schema in [
-                "age INT, name STRING",
-                StructType(
-                    [
-                        StructField("age", IntegerType()),
-                        StructField("name", StringType()),
-                    ]
-                ),
-            ]:
-                self.assert_eq(
-                    self.connect.read.json(path=d, schema=schema).toPandas(),
-                    self.spark.read.json(path=d, schema=schema).toPandas(),
-                )
-
-            self.assert_eq(
-                self.connect.read.json(path=d, primitivesAsString=True).toPandas(),
-                self.spark.read.json(path=d, primitivesAsString=True).toPandas(),
-            )
-
-    def test_xml(self):
-        tmpPath = tempfile.mkdtemp()
-        shutil.rmtree(tmpPath)
-        xsdPath = tempfile.mkdtemp()
-        xsdString = """<?xml version="1.0" encoding="UTF-8" ?>
-          <xs:schema xmlns:xs="http://www.w3.org/2001/XMLSchema">
-            <xs:element name="person">
-              <xs:complexType>
-                <xs:sequence>
-                  <xs:element name="name" type="xs:string" />
-                  <xs:element name="age" type="xs:long" />
-                </xs:sequence>
-              </xs:complexType>
-            </xs:element>
-          </xs:schema>"""
-        try:
-            xsdFile = os.path.join(xsdPath, "people.xsd")
-            with open(xsdFile, "w") as f:
-                _ = f.write(xsdString)
-            df = self.spark.createDataFrame([("Hyukjin", 100), ("Aria", 101), ("Arin", 102)]).toDF(
-                "name", "age"
-            )
-            df.write.xml(tmpPath, rootTag="people", rowTag="person")
-            people = self.spark.read.xml(tmpPath, rowTag="person", rowValidationXSDPath=xsdFile)
-            peopleConnect = self.connect.read.xml(
-                tmpPath, rowTag="person", rowValidationXSDPath=xsdFile
-            )
-            self.assert_eq(people.toPandas(), peopleConnect.toPandas())
-            expected = [
-                Row(age=100, name="Hyukjin"),
-                Row(age=101, name="Aria"),
-                Row(age=102, name="Arin"),
-            ]
-            expectedSchema = StructType(
-                [StructField("age", LongType(), True), StructField("name", StringType(), True)]
-            )
-
-            self.assertEqual(people.sort("age").collect(), expected)
-            self.assertEqual(people.schema, expectedSchema)
-
-            for schema in [
-                "age INT, name STRING",
-                expectedSchema,
-            ]:
-                people = self.spark.read.xml(
-                    tmpPath, rowTag="person", rowValidationXSDPath=xsdFile, schema=schema
-                )
-                peopleConnect = self.connect.read.xml(
-                    tmpPath, rowTag="person", rowValidationXSDPath=xsdFile, schema=schema
-                )
-                self.assert_eq(people.toPandas(), peopleConnect.toPandas())
-
-        finally:
-            shutil.rmtree(tmpPath)
-            shutil.rmtree(xsdPath)
-
-    def test_parquet(self):
-        # SPARK-41445: Implement DataFrameReader.parquet
-        with tempfile.TemporaryDirectory(prefix="test_parquet") as d:
-            # Write a DataFrame into a JSON file
-            self.spark.createDataFrame([{"age": 100, "name": "Hyukjin Kwon"}]).write.mode(
-                "overwrite"
-            ).format("parquet").save(d)
-            # Read the Parquet file as a DataFrame.
-            self.assert_eq(
-                self.connect.read.parquet(d).toPandas(), self.spark.read.parquet(d).toPandas()
-            )
-
-    def test_text(self):
-        # SPARK-41849: Implement DataFrameReader.text
-        with tempfile.TemporaryDirectory(prefix="test_text") as d:
-            # Write a DataFrame into a text file
-            self.spark.createDataFrame(
-                [{"name": "Sandeep Singh"}, {"name": "Hyukjin Kwon"}]
-            ).write.mode("overwrite").format("text").save(d)
-            # Read the text file as a DataFrame.
-            self.assert_eq(self.connect.read.text(d).toPandas(), self.spark.read.text(d).toPandas())
-
-    def test_csv(self):
-        # SPARK-42011: Implement DataFrameReader.csv
-        with tempfile.TemporaryDirectory(prefix="test_csv") as d:
-            # Write a DataFrame into a text file
-            self.spark.createDataFrame(
-                [{"name": "Sandeep Singh"}, {"name": "Hyukjin Kwon"}]
-            ).write.mode("overwrite").format("csv").save(d)
-            # Read the text file as a DataFrame.
-            self.assert_eq(self.connect.read.csv(d).toPandas(), self.spark.read.csv(d).toPandas())
-
-    def test_multi_paths(self):
-        # SPARK-42041: DataFrameReader should support list of paths
-
-        with tempfile.TemporaryDirectory(prefix="test_multi_paths1") as d:
-            text_files = []
-            for i in range(0, 3):
-                text_file = f"{d}/text-{i}.text"
-                shutil.copyfile("python/test_support/sql/text-test.txt", text_file)
-                text_files.append(text_file)
-
-            self.assertEqual(
-                self.connect.read.text(text_files).collect(),
-                self.spark.read.text(text_files).collect(),
-            )
-
-        with tempfile.TemporaryDirectory(prefix="test_multi_paths2") as d:
-            json_files = []
-            for i in range(0, 5):
-                json_file = f"{d}/json-{i}.json"
-                shutil.copyfile("python/test_support/sql/people.json", json_file)
-                json_files.append(json_file)
-
-            self.assertEqual(
-                self.connect.read.json(json_files).collect(),
-                self.spark.read.json(json_files).collect(),
-            )
-
-    def test_orc(self):
-        # SPARK-42012: Implement DataFrameReader.orc
-        with tempfile.TemporaryDirectory(prefix="test_orc") as d:
-            # Write a DataFrame into a text file
-            self.spark.createDataFrame(
-                [{"name": "Sandeep Singh"}, {"name": "Hyukjin Kwon"}]
-            ).write.mode("overwrite").format("orc").save(d)
-            # Read the text file as a DataFrame.
-            self.assert_eq(self.connect.read.orc(d).toPandas(), self.spark.read.orc(d).toPandas())
-
     def test_join_condition_column_list_columns(self):
         left_connect_df = self.connect.read.table(self.tbl_name)
         right_connect_df = self.connect.read.table(self.tbl_name2)
@@ -586,52 +418,6 @@ class SparkConnectBasicTests(SparkConnectSQLTestCase):
         ):
             cdf1.join(cdf1).select(cdf1["*"]).schema
 
-    def test_collect(self):
-        cdf = self.connect.read.table(self.tbl_name)
-        sdf = self.spark.read.table(self.tbl_name)
-
-        data = cdf.limit(10).collect()
-        self.assertEqual(len(data), 10)
-        # Check Row has schema column names.
-        self.assertTrue("name" in data[0])
-        self.assertTrue("id" in data[0])
-
-        cdf = cdf.select(
-            CF.log("id"), CF.log("id"), CF.struct("id", "name"), CF.struct("id", "name")
-        ).limit(10)
-        sdf = sdf.select(
-            SF.log("id"), SF.log("id"), SF.struct("id", "name"), SF.struct("id", "name")
-        ).limit(10)
-
-        self.assertEqual(
-            cdf.collect(),
-            sdf.collect(),
-        )
-
-    def test_collect_timestamp(self):
-        query = """
-            SELECT * FROM VALUES
-            (TIMESTAMP('2022-12-25 10:30:00'), 1),
-            (TIMESTAMP('2022-12-25 10:31:00'), 2),
-            (TIMESTAMP('2022-12-25 10:32:00'), 1),
-            (TIMESTAMP('2022-12-25 10:33:00'), 2),
-            (TIMESTAMP('2022-12-26 09:30:00'), 1),
-            (TIMESTAMP('2022-12-26 09:35:00'), 3)
-            AS tab(date, val)
-            """
-
-        cdf = self.connect.sql(query)
-        sdf = self.spark.sql(query)
-
-        self.assertEqual(cdf.schema, sdf.schema)
-
-        self.assertEqual(cdf.collect(), sdf.collect())
-
-        self.assertEqual(
-            cdf.select(CF.date_trunc("year", cdf.date).alias("year")).collect(),
-            sdf.select(SF.date_trunc("year", sdf.date).alias("year")).collect(),
-        )
-
     def test_with_columns_renamed(self):
         # SPARK-41312: test DataFrame.withColumnsRenamed()
         self.assertEqual(
@@ -646,485 +432,6 @@ class SparkConnectBasicTests(SparkConnectSQLTestCase):
             .withColumnsRenamed({"id": "id_new", "name": "name_new"})
             .schema,
         )
-
-    def test_with_local_data(self):
-        """SPARK-41114: Test creating a dataframe using local data"""
-        pdf = pd.DataFrame({"a": [1, 2, 3], "b": ["a", "b", "c"]})
-        df = self.connect.createDataFrame(pdf)
-        rows = df.filter(df.a == CF.lit(3)).collect()
-        self.assertTrue(len(rows) == 1)
-        self.assertEqual(rows[0][0], 3)
-        self.assertEqual(rows[0][1], "c")
-
-        # Check correct behavior for empty DataFrame
-        pdf = pd.DataFrame({"a": []})
-        with self.assertRaises(ValueError):
-            self.connect.createDataFrame(pdf)
-
-    def test_with_local_ndarray(self):
-        """SPARK-41446: Test creating a dataframe using local list"""
-        data = np.array([[1, 2, 3, 4], [5, 6, 7, 8]])
-
-        sdf = self.spark.createDataFrame(data)
-        cdf = self.connect.createDataFrame(data)
-        self.assertEqual(sdf.schema, cdf.schema)
-        self.assert_eq(sdf.toPandas(), cdf.toPandas())
-
-        for schema in [
-            StructType(
-                [
-                    StructField("col1", IntegerType(), True),
-                    StructField("col2", IntegerType(), True),
-                    StructField("col3", IntegerType(), True),
-                    StructField("col4", IntegerType(), True),
-                ]
-            ),
-            "struct<col1 int, col2 int, col3 int, col4 int>",
-            "col1 int, col2 int, col3 int, col4 int",
-            "col1 int, col2 long, col3 string, col4 long",
-            "col1 int, col2 string, col3 short, col4 long",
-            ["a", "b", "c", "d"],
-            ("x1", "x2", "x3", "x4"),
-        ]:
-            with self.subTest(schema=schema):
-                sdf = self.spark.createDataFrame(data, schema=schema)
-                cdf = self.connect.createDataFrame(data, schema=schema)
-
-                self.assertEqual(sdf.schema, cdf.schema)
-                self.assert_eq(sdf.toPandas(), cdf.toPandas())
-
-        with self.assertRaises(PySparkValueError) as pe:
-            self.connect.createDataFrame(data, ["a", "b", "c", "d", "e"])
-
-        self.check_error(
-            exception=pe.exception,
-            error_class="AXIS_LENGTH_MISMATCH",
-            message_parameters={"expected_length": "5", "actual_length": "4"},
-        )
-
-        with self.assertRaises(ParseException):
-            self.connect.createDataFrame(data, "col1 magic_type, col2 int, col3 int, col4 int")
-
-        with self.assertRaises(PySparkValueError) as pe:
-            self.connect.createDataFrame(data, "col1 int, col2 int, col3 int")
-
-        self.check_error(
-            exception=pe.exception,
-            error_class="AXIS_LENGTH_MISMATCH",
-            message_parameters={"expected_length": "3", "actual_length": "4"},
-        )
-
-        # test 1 dim ndarray
-        data = np.array([1.0, 2.0, np.nan, 3.0, 4.0, float("NaN"), 5.0])
-        self.assertEqual(data.ndim, 1)
-
-        sdf = self.spark.createDataFrame(data)
-        cdf = self.connect.createDataFrame(data)
-        self.assertEqual(sdf.schema, cdf.schema)
-        self.assert_eq(sdf.toPandas(), cdf.toPandas())
-
-    def test_with_local_list(self):
-        """SPARK-41446: Test creating a dataframe using local list"""
-        data = [[1, 2, 3, 4]]
-
-        sdf = self.spark.createDataFrame(data)
-        cdf = self.connect.createDataFrame(data)
-        self.assertEqual(sdf.schema, cdf.schema)
-        self.assert_eq(sdf.toPandas(), cdf.toPandas())
-
-        for schema in [
-            "struct<col1 int, col2 int, col3 int, col4 int>",
-            "col1 int, col2 int, col3 int, col4 int",
-            "col1 int, col2 long, col3 string, col4 long",
-            "col1 int, col2 string, col3 short, col4 long",
-            ["a", "b", "c", "d"],
-            ("x1", "x2", "x3", "x4"),
-        ]:
-            sdf = self.spark.createDataFrame(data, schema=schema)
-            cdf = self.connect.createDataFrame(data, schema=schema)
-
-            self.assertEqual(sdf.schema, cdf.schema)
-            self.assert_eq(sdf.toPandas(), cdf.toPandas())
-
-        with self.assertRaises(PySparkValueError) as pe:
-            self.connect.createDataFrame(data, ["a", "b", "c", "d", "e"])
-
-        self.check_error(
-            exception=pe.exception,
-            error_class="AXIS_LENGTH_MISMATCH",
-            message_parameters={"expected_length": "5", "actual_length": "4"},
-        )
-
-        with self.assertRaises(ParseException):
-            self.connect.createDataFrame(data, "col1 magic_type, col2 int, col3 int, col4 int")
-
-        with self.assertRaises(PySparkValueError) as pe:
-            self.connect.createDataFrame(data, "col1 int, col2 int, col3 int")
-
-        self.check_error(
-            exception=pe.exception,
-            error_class="AXIS_LENGTH_MISMATCH",
-            message_parameters={"expected_length": "3", "actual_length": "4"},
-        )
-
-    def test_with_local_rows(self):
-        # SPARK-41789, SPARK-41810: Test creating a dataframe with list of rows and dictionaries
-        rows = [
-            Row(course="dotNET", year=2012, earnings=10000),
-            Row(course="Java", year=2012, earnings=20000),
-            Row(course="dotNET", year=2012, earnings=5000),
-            Row(course="dotNET", year=2013, earnings=48000),
-            Row(course="Java", year=2013, earnings=30000),
-            Row(course="Scala", year=2022, earnings=None),
-        ]
-        dicts = [row.asDict() for row in rows]
-
-        for data in [rows, dicts]:
-            sdf = self.spark.createDataFrame(data)
-            cdf = self.connect.createDataFrame(data)
-
-            self.assertEqual(sdf.schema, cdf.schema)
-            self.assert_eq(sdf.toPandas(), cdf.toPandas())
-
-            # test with rename
-            sdf = self.spark.createDataFrame(data, schema=["a", "b", "c"])
-            cdf = self.connect.createDataFrame(data, schema=["a", "b", "c"])
-
-            self.assertEqual(sdf.schema, cdf.schema)
-            self.assert_eq(sdf.toPandas(), cdf.toPandas())
-
-    def test_streaming_local_relation(self):
-        threshold_conf = "spark.sql.session.localRelationCacheThreshold"
-        old_threshold = self.connect.conf.get(threshold_conf)
-        threshold = 1024 * 1024
-        self.connect.conf.set(threshold_conf, threshold)
-        try:
-            suffix = "abcdef"
-            letters = string.ascii_lowercase
-            str = "".join(random.choice(letters) for i in range(threshold)) + suffix
-            data = [[0, str], [1, str]]
-            for i in range(0, 2):
-                cdf = self.connect.createDataFrame(data, ["a", "b"])
-                self.assert_eq(cdf.count(), len(data))
-                self.assert_eq(cdf.filter(f"endsWith(b, '{suffix}')").isEmpty(), False)
-        finally:
-            self.connect.conf.set(threshold_conf, old_threshold)
-
-    def test_with_atom_type(self):
-        for data in [[(1), (2), (3)], [1, 2, 3]]:
-            for schema in ["long", "int", "short"]:
-                sdf = self.spark.createDataFrame(data, schema=schema)
-                cdf = self.connect.createDataFrame(data, schema=schema)
-
-                self.assertEqual(sdf.schema, cdf.schema)
-                self.assert_eq(sdf.toPandas(), cdf.toPandas())
-
-    def test_with_none_and_nan(self):
-        # SPARK-41855: make createDataFrame support None and NaN
-        # SPARK-41814: test with eqNullSafe
-        data1 = [Row(id=1, value=float("NaN")), Row(id=2, value=42.0), Row(id=3, value=None)]
-        data2 = [Row(id=1, value=np.nan), Row(id=2, value=42.0), Row(id=3, value=None)]
-        data3 = [
-            {"id": 1, "value": float("NaN")},
-            {"id": 2, "value": 42.0},
-            {"id": 3, "value": None},
-        ]
-        data4 = [{"id": 1, "value": np.nan}, {"id": 2, "value": 42.0}, {"id": 3, "value": None}]
-        data5 = [(1, float("NaN")), (2, 42.0), (3, None)]
-        data6 = [(1, np.nan), (2, 42.0), (3, None)]
-        data7 = np.array([[1, float("NaN")], [2, 42.0], [3, None]])
-        data8 = np.array([[1, np.nan], [2, 42.0], [3, None]])
-
-        # +---+-----+
-        # | id|value|
-        # +---+-----+
-        # |  1|  NaN|
-        # |  2| 42.0|
-        # |  3| NULL|
-        # +---+-----+
-
-        for data in [data1, data2, data3, data4, data5, data6, data7, data8]:
-            if isinstance(data[0], (Row, dict)):
-                # data1, data2, data3, data4
-                cdf = self.connect.createDataFrame(data)
-                sdf = self.spark.createDataFrame(data)
-            else:
-                # data5, data6, data7, data8
-                cdf = self.connect.createDataFrame(data, schema=["id", "value"])
-                sdf = self.spark.createDataFrame(data, schema=["id", "value"])
-
-            self.assert_eq(cdf.toPandas(), sdf.toPandas())
-
-            self.assert_eq(
-                cdf.select(
-                    cdf["value"].eqNullSafe(None),
-                    cdf["value"].eqNullSafe(float("NaN")),
-                    cdf["value"].eqNullSafe(42.0),
-                ).toPandas(),
-                sdf.select(
-                    sdf["value"].eqNullSafe(None),
-                    sdf["value"].eqNullSafe(float("NaN")),
-                    sdf["value"].eqNullSafe(42.0),
-                ).toPandas(),
-            )
-
-        # SPARK-41851: test with nanvl
-        data = [(1.0, float("nan")), (float("nan"), 2.0)]
-
-        cdf = self.connect.createDataFrame(data, ("a", "b"))
-        sdf = self.spark.createDataFrame(data, ("a", "b"))
-
-        self.assert_eq(cdf.toPandas(), sdf.toPandas())
-
-        self.assert_eq(
-            cdf.select(
-                CF.nanvl("a", "b").alias("r1"), CF.nanvl(cdf.a, cdf.b).alias("r2")
-            ).toPandas(),
-            sdf.select(
-                SF.nanvl("a", "b").alias("r1"), SF.nanvl(sdf.a, sdf.b).alias("r2")
-            ).toPandas(),
-        )
-
-        # SPARK-41852: test with pmod
-        data = [
-            (1.0, float("nan")),
-            (float("nan"), 2.0),
-            (10.0, 3.0),
-            (float("nan"), float("nan")),
-            (-3.0, 4.0),
-            (-10.0, 3.0),
-            (-5.0, -6.0),
-            (7.0, -8.0),
-            (1.0, 2.0),
-        ]
-
-        cdf = self.connect.createDataFrame(data, ("a", "b"))
-        sdf = self.spark.createDataFrame(data, ("a", "b"))
-
-        self.assert_eq(cdf.toPandas(), sdf.toPandas())
-
-        self.assert_eq(
-            cdf.select(CF.pmod("a", "b")).toPandas(),
-            sdf.select(SF.pmod("a", "b")).toPandas(),
-        )
-
-    def test_cast_with_ddl(self):
-        data = [Row(date=datetime.date(2021, 12, 27), add=2)]
-
-        cdf = self.connect.createDataFrame(data, "date date, add integer")
-        sdf = self.spark.createDataFrame(data, "date date, add integer")
-
-        self.assertEqual(cdf.schema, sdf.schema)
-
-    def test_create_empty_df(self):
-        for schema in [
-            "STRING",
-            "x STRING",
-            "x STRING, y INTEGER",
-            StringType(),
-            StructType(
-                [
-                    StructField("x", StringType(), True),
-                    StructField("y", IntegerType(), True),
-                ]
-            ),
-        ]:
-            cdf = self.connect.createDataFrame(data=[], schema=schema)
-            sdf = self.spark.createDataFrame(data=[], schema=schema)
-
-            self.assert_eq(cdf.toPandas(), sdf.toPandas())
-
-        # check error
-        with self.assertRaises(PySparkValueError) as pe:
-            self.connect.createDataFrame(data=[])
-
-        self.check_error(
-            exception=pe.exception,
-            error_class="CANNOT_INFER_EMPTY_SCHEMA",
-            message_parameters={},
-        )
-
-    def test_create_dataframe_from_arrays(self):
-        # SPARK-42021: createDataFrame support array.array
-        data1 = [Row(a=1, b=array.array("i", [1, 2, 3]), c=array.array("d", [4, 5, 6]))]
-        data2 = [(array.array("d", [1, 2, 3]), 2, "3")]
-        data3 = [{"a": 1, "b": array.array("i", [1, 2, 3])}]
-
-        for data in [data1, data2, data3]:
-            cdf = self.connect.createDataFrame(data)
-            sdf = self.spark.createDataFrame(data)
-
-            self.assertEqual(cdf.schema, sdf.schema)
-            self.assertEqual(cdf.collect(), sdf.collect())
-
-    def test_timestampe_create_from_rows(self):
-        data = [(datetime.datetime(2016, 3, 11, 9, 0, 7), 1)]
-
-        cdf = self.connect.createDataFrame(data, ["date", "val"])
-        sdf = self.spark.createDataFrame(data, ["date", "val"])
-
-        self.assertEqual(cdf.schema, sdf.schema)
-        self.assertEqual(cdf.collect(), sdf.collect())
-
-    def test_create_dataframe_with_coercion(self):
-        data1 = [[1.33, 1], ["2.1", 1]]
-        data2 = [[True, 1], ["false", 1]]
-
-        for data in [data1, data2]:
-            cdf = self.connect.createDataFrame(data, ["a", "b"])
-            sdf = self.spark.createDataFrame(data, ["a", "b"])
-
-            self.assertEqual(cdf.schema, sdf.schema)
-            self.assertEqual(cdf.collect(), sdf.collect())
-
-    def test_nested_type_create_from_rows(self):
-        data1 = [Row(a=1, b=Row(c=2, d=Row(e=3, f=Row(g=4, h=Row(i=5)))))]
-        # root
-        # |-- a: long (nullable = true)
-        # |-- b: struct (nullable = true)
-        # |    |-- c: long (nullable = true)
-        # |    |-- d: struct (nullable = true)
-        # |    |    |-- e: long (nullable = true)
-        # |    |    |-- f: struct (nullable = true)
-        # |    |    |    |-- g: long (nullable = true)
-        # |    |    |    |-- h: struct (nullable = true)
-        # |    |    |    |    |-- i: long (nullable = true)
-
-        data2 = [
-            (
-                1,
-                "a",
-                Row(
-                    a=1,
-                    b=[1, 2, 3],
-                    c={"a": "b"},
-                    d=Row(x=1, y="y", z=Row(o=1, p=2, q=Row(g=1.5))),
-                ),
-            )
-        ]
-        # root
-        # |-- _1: long (nullable = true)
-        # |-- _2: string (nullable = true)
-        # |-- _3: struct (nullable = true)
-        # |    |-- a: long (nullable = true)
-        # |    |-- b: array (nullable = true)
-        # |    |    |-- element: long (containsNull = true)
-        # |    |-- c: map (nullable = true)
-        # |    |    |-- key: string
-        # |    |    |-- value: string (valueContainsNull = true)
-        # |    |-- d: struct (nullable = true)
-        # |    |    |-- x: long (nullable = true)
-        # |    |    |-- y: string (nullable = true)
-        # |    |    |-- z: struct (nullable = true)
-        # |    |    |    |-- o: long (nullable = true)
-        # |    |    |    |-- p: long (nullable = true)
-        # |    |    |    |-- q: struct (nullable = true)
-        # |    |    |    |    |-- g: double (nullable = true)
-
-        data3 = [
-            Row(
-                a=1,
-                b=[1, 2, 3],
-                c={"a": "b"},
-                d=Row(x=1, y="y", z=Row(1, 2, 3)),
-                e=list("hello connect"),
-            )
-        ]
-        # root
-        # |-- a: long (nullable = true)
-        # |-- b: array (nullable = true)
-        # |    |-- element: long (containsNull = true)
-        # |-- c: map (nullable = true)
-        # |    |-- key: string
-        # |    |-- value: string (valueContainsNull = true)
-        # |-- d: struct (nullable = true)
-        # |    |-- x: long (nullable = true)
-        # |    |-- y: string (nullable = true)
-        # |    |-- z: struct (nullable = true)
-        # |    |    |-- _1: long (nullable = true)
-        # |    |    |-- _2: long (nullable = true)
-        # |    |    |-- _3: long (nullable = true)
-        # |-- e: array (nullable = true)
-        # |    |-- element: string (containsNull = true)
-
-        data4 = [
-            {
-                "a": 1,
-                "b": Row(x=1, y=Row(z=2)),
-                "c": {"x": -1, "y": 2},
-                "d": [1, 2, 3, 4, 5],
-            }
-        ]
-        # root
-        # |-- a: long (nullable = true)
-        # |-- b: struct (nullable = true)
-        # |    |-- x: long (nullable = true)
-        # |    |-- y: struct (nullable = true)
-        # |    |    |-- z: long (nullable = true)
-        # |-- c: map (nullable = true)
-        # |    |-- key: string
-        # |    |-- value: long (valueContainsNull = true)
-        # |-- d: array (nullable = true)
-        # |    |-- element: long (containsNull = true)
-
-        data5 = [
-            {
-                "a": [Row(x=1, y="2"), Row(x=-1, y="-2")],
-                "b": [[1, 2, 3], [4, 5], [6]],
-                "c": {3: {4: {5: 6}}, 7: {8: {9: 0}}},
-            }
-        ]
-        # root
-        # |-- a: array (nullable = true)
-        # |    |-- element: struct (containsNull = true)
-        # |    |    |-- x: long (nullable = true)
-        # |    |    |-- y: string (nullable = true)
-        # |-- b: array (nullable = true)
-        # |    |-- element: array (containsNull = true)
-        # |    |    |-- element: long (containsNull = true)
-        # |-- c: map (nullable = true)
-        # |    |-- key: long
-        # |    |-- value: map (valueContainsNull = true)
-        # |    |    |-- key: long
-        # |    |    |-- value: map (valueContainsNull = true)
-        # |    |    |    |-- key: long
-        # |    |    |    |-- value: long (valueContainsNull = true)
-
-        for data in [data1, data2, data3, data4, data5]:
-            with self.subTest(data=data):
-                cdf = self.connect.createDataFrame(data)
-                sdf = self.spark.createDataFrame(data)
-
-                self.assertEqual(cdf.schema, sdf.schema)
-                self.assertEqual(cdf.collect(), sdf.collect())
-
-    def test_create_df_from_objects(self):
-        data = [MyObject(1, "1"), MyObject(2, "2")]
-
-        # +---+-----+
-        # |key|value|
-        # +---+-----+
-        # |  1|    1|
-        # |  2|    2|
-        # +---+-----+
-
-        cdf = self.connect.createDataFrame(data)
-        sdf = self.spark.createDataFrame(data)
-
-        self.assertEqual(cdf.schema, sdf.schema)
-        self.assertEqual(cdf.collect(), sdf.collect())
-
-    def test_create_df_nullability(self):
-        data = [("asd", None)]
-        schema = StructType(
-            [
-                StructField("name", StringType(), nullable=True),
-                StructField("age", IntegerType(), nullable=False),
-            ]
-        )
-
-        with self.assertRaises(PySparkValueError):
-            self.spark.createDataFrame(data, schema)
 
     def test_simple_explain_string(self):
         df = self.connect.read.table(self.tbl_name).limit(10)
@@ -1401,15 +708,6 @@ class SparkConnectBasicTests(SparkConnectSQLTestCase):
                 message_parameters={"arg_name": "args", "arg_type": "set"},
             )
 
-    def test_head(self):
-        # SPARK-41002: test `head` API in Python Client
-        df = self.connect.read.table(self.tbl_name)
-        self.assertIsNotNone(len(df.head()))
-        self.assertIsNotNone(len(df.head(1)))
-        self.assertIsNotNone(len(df.head(5)))
-        df2 = self.connect.read.table(self.tbl_name_empty)
-        self.assertIsNone(df2.head())
-
     def test_deduplicate(self):
         # SPARK-41326: test distinct and dropDuplicates.
         df = self.connect.read.table(self.tbl_name)
@@ -1427,20 +725,6 @@ class SparkConnectBasicTests(SparkConnectSQLTestCase):
             "dropDuplicatesWithinWatermark is not supported with batch DataFrames/DataSets",
         ):
             df.dropDuplicatesWithinWatermark().toPandas()
-
-    def test_first(self):
-        # SPARK-41002: test `first` API in Python Client
-        df = self.connect.read.table(self.tbl_name)
-        self.assertIsNotNone(len(df.first()))
-        df2 = self.connect.read.table(self.tbl_name_empty)
-        self.assertIsNone(df2.first())
-
-    def test_take(self) -> None:
-        # SPARK-41002: test `take` API in Python Client
-        df = self.connect.read.table(self.tbl_name)
-        self.assertEqual(5, len(df.take(5)))
-        df2 = self.connect.read.table(self.tbl_name_empty)
-        self.assertEqual(0, len(df2.take(5)))
 
     def test_drop(self):
         # SPARK-41169: test drop
@@ -1561,89 +845,6 @@ class SparkConnectBasicTests(SparkConnectSQLTestCase):
             # Test when creating a view which is already exists but
             with self.assertRaises(AnalysisException):
                 self.connect.sql("SELECT 1 AS X LIMIT 0").createTempView("view_local_temp")
-
-    def test_to_pandas(self):
-        # SPARK-41005: Test to pandas
-        query = """
-            SELECT * FROM VALUES
-            (false, 1, NULL),
-            (false, NULL, float(2.0)),
-            (NULL, 3, float(3.0))
-            AS tab(a, b, c)
-            """
-
-        self.assert_eq(
-            self.connect.sql(query).toPandas(),
-            self.spark.sql(query).toPandas(),
-        )
-
-        query = """
-            SELECT * FROM VALUES
-            (1, 1, NULL),
-            (2, NULL, float(2.0)),
-            (3, 3, float(3.0))
-            AS tab(a, b, c)
-            """
-
-        self.assert_eq(
-            self.connect.sql(query).toPandas(),
-            self.spark.sql(query).toPandas(),
-        )
-
-        query = """
-            SELECT * FROM VALUES
-            (double(1.0), 1, "1"),
-            (NULL, NULL, NULL),
-            (double(2.0), 3, "3")
-            AS tab(a, b, c)
-            """
-
-        self.assert_eq(
-            self.connect.sql(query).toPandas(),
-            self.spark.sql(query).toPandas(),
-        )
-
-        query = """
-            SELECT * FROM VALUES
-            (float(1.0), double(1.0), 1, "1"),
-            (float(2.0), double(2.0), 2, "2"),
-            (float(3.0), double(3.0), 3, "3")
-            AS tab(a, b, c, d)
-            """
-
-        self.assert_eq(
-            self.connect.sql(query).toPandas(),
-            self.spark.sql(query).toPandas(),
-        )
-
-    def test_create_dataframe_from_pandas_with_ns_timestamp(self):
-        """Truncate the timestamps for nanoseconds."""
-        from datetime import datetime, timezone, timedelta
-        from pandas import Timestamp
-        import pandas as pd
-
-        pdf = pd.DataFrame(
-            {
-                "naive": [datetime(2019, 1, 1, 0)],
-                "aware": [
-                    Timestamp(
-                        year=2019, month=1, day=1, nanosecond=500, tz=timezone(timedelta(hours=-8))
-                    )
-                ],
-            }
-        )
-
-        with self.sql_conf({"spark.sql.execution.arrow.pyspark.enabled": False}):
-            self.assertEqual(
-                self.connect.createDataFrame(pdf).collect(),
-                self.spark.createDataFrame(pdf).collect(),
-            )
-
-        with self.sql_conf({"spark.sql.execution.arrow.pyspark.enabled": True}):
-            self.assertEqual(
-                self.connect.createDataFrame(pdf).collect(),
-                self.spark.createDataFrame(pdf).collect(),
-            )
 
     def test_select_expr(self):
         # SPARK-41201: test selectExpr API.
@@ -2327,37 +1528,6 @@ class SparkConnectBasicTests(SparkConnectSQLTestCase):
             message_parameters={"explain_mode": "unknown"},
         )
 
-    def test_simple_datasource_read(self) -> None:
-        writeDf = self.df_text
-        tmpPath = tempfile.mkdtemp()
-        shutil.rmtree(tmpPath)
-        writeDf.write.text(tmpPath)
-
-        for schema in [
-            "id STRING",
-            StructType([StructField("id", StringType())]),
-        ]:
-            readDf = self.connect.read.format("text").schema(schema).load(path=tmpPath)
-            expectResult = writeDf.collect()
-            pandasResult = readDf.toPandas()
-            if pandasResult is None:
-                self.assertTrue(False, "Empty pandas dataframe")
-            else:
-                actualResult = pandasResult.values.tolist()
-                self.assertEqual(len(expectResult), len(actualResult))
-
-    def test_simple_read_without_schema(self) -> None:
-        """SPARK-41300: Schema not set when reading CSV."""
-        writeDf = self.df_text
-        tmpPath = tempfile.mkdtemp()
-        shutil.rmtree(tmpPath)
-        writeDf.write.csv(tmpPath, header=True)
-
-        readDf = self.connect.read.format("csv").option("header", True).load(path=tmpPath)
-        expectResult = set(writeDf.collect())
-        pandasResult = set(readDf.collect())
-        self.assertEqual(expectResult, pandasResult)
-
     def test_count(self) -> None:
         # SPARK-41308: test count() API.
         self.assertEqual(
@@ -2473,50 +1643,6 @@ class SparkConnectBasicTests(SparkConnectSQLTestCase):
             ndf1.subtract(ndf2).toPandas(),
             df1.subtract(df2).toPandas(),
         )
-
-    def test_write_operations(self):
-        with tempfile.TemporaryDirectory(prefix="test_write_operations") as d:
-            df = self.connect.range(50)
-            df.write.mode("overwrite").format("csv").save(d)
-
-            ndf = self.connect.read.schema("id int").load(d, format="csv")
-            self.assertEqual(50, len(ndf.collect()))
-            cd = ndf.collect()
-            self.assertEqual(set(df.collect()), set(cd))
-
-        with tempfile.TemporaryDirectory(prefix="test_write_operations") as d:
-            df = self.connect.range(50)
-            df.write.mode("overwrite").csv(d, lineSep="|")
-
-            ndf = self.connect.read.schema("id int").load(d, format="csv", lineSep="|")
-            self.assertEqual(set(df.collect()), set(ndf.collect()))
-
-        df = self.connect.range(50)
-        df.write.format("parquet").saveAsTable("parquet_test")
-
-        ndf = self.connect.read.table("parquet_test")
-        self.assertEqual(set(df.collect()), set(ndf.collect()))
-
-    def test_writeTo_operations(self):
-        # SPARK-42002: Implement DataFrameWriterV2
-        import datetime
-        from pyspark.sql.connect.functions import col, years, months, days, hours, bucket
-
-        df = self.connect.createDataFrame(
-            [(1, datetime.datetime(2000, 1, 1), "foo")], ("id", "ts", "value")
-        )
-        writer = df.writeTo("table1")
-        self.assertIsInstance(writer.option("property", "value"), DataFrameWriterV2)
-        self.assertIsInstance(writer.options(property="value"), DataFrameWriterV2)
-        self.assertIsInstance(writer.using("source"), DataFrameWriterV2)
-        self.assertIsInstance(writer.partitionedBy(col("id")), DataFrameWriterV2)
-        self.assertIsInstance(writer.tableProperty("foo", "bar"), DataFrameWriterV2)
-        self.assertIsInstance(writer.partitionedBy(years("ts")), DataFrameWriterV2)
-        self.assertIsInstance(writer.partitionedBy(months("ts")), DataFrameWriterV2)
-        self.assertIsInstance(writer.partitionedBy(days("ts")), DataFrameWriterV2)
-        self.assertIsInstance(writer.partitionedBy(hours("ts")), DataFrameWriterV2)
-        self.assertIsInstance(writer.partitionedBy(bucket(11, "id")), DataFrameWriterV2)
-        self.assertIsInstance(writer.partitionedBy(bucket(3, "id"), hours("ts")), DataFrameWriterV2)
 
     def test_agg_with_avg(self):
         # SPARK-41325: groupby.avg()
@@ -2960,148 +2086,6 @@ class SparkConnectBasicTests(SparkConnectSQLTestCase):
             },
         )
 
-    def test_collect_nested_type(self):
-        query = """
-            SELECT * FROM VALUES
-            (1, 4, 0, 8, true, true, ARRAY(1, NULL, 3), MAP(1, 2, 3, 4)),
-            (2, 5, -1, NULL, false, NULL, ARRAY(1, 3), MAP(1, NULL, 3, 4)),
-            (3, 6, NULL, 0, false, NULL, ARRAY(NULL), NULL)
-            AS tab(a, b, c, d, e, f, g, h)
-            """
-
-        # +---+---+----+----+-----+----+------------+-------------------+
-        # |  a|  b|   c|   d|    e|   f|           g|                  h|
-        # +---+---+----+----+-----+----+------------+-------------------+
-        # |  1|  4|   0|   8| true|true|[1, null, 3]|   {1 -> 2, 3 -> 4}|
-        # |  2|  5|  -1|NULL|false|NULL|      [1, 3]|{1 -> null, 3 -> 4}|
-        # |  3|  6|NULL|   0|false|NULL|      [null]|               NULL|
-        # +---+---+----+----+-----+----+------------+-------------------+
-
-        cdf = self.connect.sql(query)
-        sdf = self.spark.sql(query)
-
-        # test collect array
-        # +--------------+-------------+------------+
-        # |array(a, b, c)|  array(e, f)|           g|
-        # +--------------+-------------+------------+
-        # |     [1, 4, 0]| [true, true]|[1, null, 3]|
-        # |    [2, 5, -1]|[false, null]|      [1, 3]|
-        # |  [3, 6, null]|[false, null]|      [null]|
-        # +--------------+-------------+------------+
-        self.assertEqual(
-            cdf.select(CF.array("a", "b", "c"), CF.array("e", "f"), CF.col("g")).collect(),
-            sdf.select(SF.array("a", "b", "c"), SF.array("e", "f"), SF.col("g")).collect(),
-        )
-
-        # test collect nested array
-        # +-----------------------------------+-------------------------+
-        # |array(array(a), array(b), array(c))|array(array(e), array(f))|
-        # +-----------------------------------+-------------------------+
-        # |                    [[1], [4], [0]]|         [[true], [true]]|
-        # |                   [[2], [5], [-1]]|        [[false], [null]]|
-        # |                 [[3], [6], [null]]|        [[false], [null]]|
-        # +-----------------------------------+-------------------------+
-        self.assertEqual(
-            cdf.select(
-                CF.array(CF.array("a"), CF.array("b"), CF.array("c")),
-                CF.array(CF.array("e"), CF.array("f")),
-            ).collect(),
-            sdf.select(
-                SF.array(SF.array("a"), SF.array("b"), SF.array("c")),
-                SF.array(SF.array("e"), SF.array("f")),
-            ).collect(),
-        )
-
-        # test collect array of struct, map
-        # +----------------+---------------------+
-        # |array(struct(a))|             array(h)|
-        # +----------------+---------------------+
-        # |           [{1}]|   [{1 -> 2, 3 -> 4}]|
-        # |           [{2}]|[{1 -> null, 3 -> 4}]|
-        # |           [{3}]|               [null]|
-        # +----------------+---------------------+
-        self.assertEqual(
-            cdf.select(CF.array(CF.struct("a")), CF.array("h")).collect(),
-            sdf.select(SF.array(SF.struct("a")), SF.array("h")).collect(),
-        )
-
-        # test collect map
-        # +-------------------+-------------------+
-        # |                  h|    map(a, b, b, c)|
-        # +-------------------+-------------------+
-        # |   {1 -> 2, 3 -> 4}|   {1 -> 4, 4 -> 0}|
-        # |{1 -> null, 3 -> 4}|  {2 -> 5, 5 -> -1}|
-        # |               NULL|{3 -> 6, 6 -> null}|
-        # +-------------------+-------------------+
-        self.assertEqual(
-            cdf.select(CF.col("h"), CF.create_map("a", "b", "b", "c")).collect(),
-            sdf.select(SF.col("h"), SF.create_map("a", "b", "b", "c")).collect(),
-        )
-
-        # test collect map of struct, array
-        # +-------------------+------------------------+
-        # |          map(a, g)|    map(a, struct(b, g))|
-        # +-------------------+------------------------+
-        # |{1 -> [1, null, 3]}|{1 -> {4, [1, null, 3]}}|
-        # |      {2 -> [1, 3]}|      {2 -> {5, [1, 3]}}|
-        # |      {3 -> [null]}|      {3 -> {6, [null]}}|
-        # +-------------------+------------------------+
-        self.assertEqual(
-            cdf.select(CF.create_map("a", "g"), CF.create_map("a", CF.struct("b", "g"))).collect(),
-            sdf.select(SF.create_map("a", "g"), SF.create_map("a", SF.struct("b", "g"))).collect(),
-        )
-
-        # test collect struct
-        # +------------------+--------------------------+
-        # |struct(a, b, c, d)|           struct(e, f, g)|
-        # +------------------+--------------------------+
-        # |      {1, 4, 0, 8}|{true, true, [1, null, 3]}|
-        # |  {2, 5, -1, null}|     {false, null, [1, 3]}|
-        # |   {3, 6, null, 0}|     {false, null, [null]}|
-        # +------------------+--------------------------+
-        self.assertEqual(
-            cdf.select(CF.struct("a", "b", "c", "d"), CF.struct("e", "f", "g")).collect(),
-            sdf.select(SF.struct("a", "b", "c", "d"), SF.struct("e", "f", "g")).collect(),
-        )
-
-        # test collect nested struct
-        # +------------------------------------------+--------------------------+----------------------------+ # noqa
-        # |struct(a, struct(a, struct(c, struct(d))))|struct(a, b, struct(c, d))|     struct(e, f, struct(g))| # noqa
-        # +------------------------------------------+--------------------------+----------------------------+ # noqa
-        # |                        {1, {1, {0, {8}}}}|            {1, 4, {0, 8}}|{true, true, {[1, null, 3]}}| # noqa
-        # |                    {2, {2, {-1, {null}}}}|        {2, 5, {-1, null}}|     {false, null, {[1, 3]}}| # noqa
-        # |                     {3, {3, {null, {0}}}}|         {3, 6, {null, 0}}|     {false, null, {[null]}}| # noqa
-        # +------------------------------------------+--------------------------+----------------------------+ # noqa
-        self.assertEqual(
-            cdf.select(
-                CF.struct("a", CF.struct("a", CF.struct("c", CF.struct("d")))),
-                CF.struct("a", "b", CF.struct("c", "d")),
-                CF.struct("e", "f", CF.struct("g")),
-            ).collect(),
-            sdf.select(
-                SF.struct("a", SF.struct("a", SF.struct("c", SF.struct("d")))),
-                SF.struct("a", "b", SF.struct("c", "d")),
-                SF.struct("e", "f", SF.struct("g")),
-            ).collect(),
-        )
-
-        # test collect struct containing array, map
-        # +--------------------------------------------+
-        # |  struct(a, struct(a, struct(g, struct(h))))|
-        # +--------------------------------------------+
-        # |{1, {1, {[1, null, 3], {{1 -> 2, 3 -> 4}}}}}|
-        # |   {2, {2, {[1, 3], {{1 -> null, 3 -> 4}}}}}|
-        # |                  {3, {3, {[null], {null}}}}|
-        # +--------------------------------------------+
-        self.assertEqual(
-            cdf.select(
-                CF.struct("a", CF.struct("a", CF.struct("g", CF.struct("h")))),
-            ).collect(),
-            sdf.select(
-                SF.struct("a", SF.struct("a", SF.struct("g", SF.struct("h")))),
-            ).collect(),
-        )
-
     def test_simple_udt(self):
         from pyspark.ml.linalg import MatrixUDT, VectorUDT
 
@@ -3208,142 +2192,6 @@ class SparkConnectBasicTests(SparkConnectSQLTestCase):
         self.assertEqual(
             self.connect.sql("show functions").collect(), self.spark.sql("show functions").collect()
         )
-
-    def test_schema_has_nullable(self):
-        schema_false = StructType().add("id", IntegerType(), False)
-        cdf1 = self.connect.createDataFrame([[1]], schema=schema_false)
-        sdf1 = self.spark.createDataFrame([[1]], schema=schema_false)
-        self.assertEqual(cdf1.schema, sdf1.schema)
-        self.assertEqual(cdf1.collect(), sdf1.collect())
-
-        schema_true = StructType().add("id", IntegerType(), True)
-        cdf2 = self.connect.createDataFrame([[1]], schema=schema_true)
-        sdf2 = self.spark.createDataFrame([[1]], schema=schema_true)
-        self.assertEqual(cdf2.schema, sdf2.schema)
-        self.assertEqual(cdf2.collect(), sdf2.collect())
-
-        pdf1 = cdf1.toPandas()
-        cdf3 = self.connect.createDataFrame(pdf1, cdf1.schema)
-        sdf3 = self.spark.createDataFrame(pdf1, sdf1.schema)
-        self.assertEqual(cdf3.schema, sdf3.schema)
-        self.assertEqual(cdf3.collect(), sdf3.collect())
-
-        pdf2 = cdf2.toPandas()
-        cdf4 = self.connect.createDataFrame(pdf2, cdf2.schema)
-        sdf4 = self.spark.createDataFrame(pdf2, sdf2.schema)
-        self.assertEqual(cdf4.schema, sdf4.schema)
-        self.assertEqual(cdf4.collect(), sdf4.collect())
-
-    def test_array_has_nullable(self):
-        for schemas, data in [
-            (
-                [StructType().add("arr", ArrayType(IntegerType(), False), True)],
-                [Row([1, 2]), Row([3]), Row(None)],
-            ),
-            (
-                [
-                    StructType().add("arr", ArrayType(IntegerType(), True), True),
-                    "arr array<integer>",
-                ],
-                [Row([1, None]), Row([3]), Row(None)],
-            ),
-            (
-                [StructType().add("arr", ArrayType(IntegerType(), False), False)],
-                [Row([1, 2]), Row([3])],
-            ),
-            (
-                [
-                    StructType().add("arr", ArrayType(IntegerType(), True), False),
-                    "arr array<integer> not null",
-                ],
-                [Row([1, None]), Row([3])],
-            ),
-        ]:
-            for schema in schemas:
-                with self.subTest(schema=schema):
-                    cdf = self.connect.createDataFrame(data, schema=schema)
-                    sdf = self.spark.createDataFrame(data, schema=schema)
-                    self.assertEqual(cdf.schema, sdf.schema)
-                    self.assertEqual(cdf.collect(), sdf.collect())
-
-    def test_map_has_nullable(self):
-        for schemas, data in [
-            (
-                [StructType().add("map", MapType(StringType(), IntegerType(), False), True)],
-                [Row({"a": 1, "b": 2}), Row({"a": 3}), Row(None)],
-            ),
-            (
-                [
-                    StructType().add("map", MapType(StringType(), IntegerType(), True), True),
-                    "map map<string, integer>",
-                ],
-                [Row({"a": 1, "b": None}), Row({"a": 3}), Row(None)],
-            ),
-            (
-                [StructType().add("map", MapType(StringType(), IntegerType(), False), False)],
-                [Row({"a": 1, "b": 2}), Row({"a": 3})],
-            ),
-            (
-                [
-                    StructType().add("map", MapType(StringType(), IntegerType(), True), False),
-                    "map map<string, integer> not null",
-                ],
-                [Row({"a": 1, "b": None}), Row({"a": 3})],
-            ),
-        ]:
-            for schema in schemas:
-                with self.subTest(schema=schema):
-                    cdf = self.connect.createDataFrame(data, schema=schema)
-                    sdf = self.spark.createDataFrame(data, schema=schema)
-                    self.assertEqual(cdf.schema, sdf.schema)
-                    self.assertEqual(cdf.collect(), sdf.collect())
-
-    def test_struct_has_nullable(self):
-        for schemas, data in [
-            (
-                [
-                    StructType().add("struct", StructType().add("i", IntegerType(), False), True),
-                    "struct struct<i: integer not null>",
-                ],
-                [Row(Row(1)), Row(Row(2)), Row(None)],
-            ),
-            (
-                [
-                    StructType().add("struct", StructType().add("i", IntegerType(), True), True),
-                    "struct struct<i: integer>",
-                ],
-                [Row(Row(1)), Row(Row(2)), Row(Row(None)), Row(None)],
-            ),
-            (
-                [
-                    StructType().add("struct", StructType().add("i", IntegerType(), False), False),
-                    "struct struct<i: integer not null> not null",
-                ],
-                [Row(Row(1)), Row(Row(2))],
-            ),
-            (
-                [
-                    StructType().add("struct", StructType().add("i", IntegerType(), True), False),
-                    "struct struct<i: integer> not null",
-                ],
-                [Row(Row(1)), Row(Row(2)), Row(Row(None))],
-            ),
-        ]:
-            for schema in schemas:
-                with self.subTest(schema=schema):
-                    cdf = self.connect.createDataFrame(data, schema=schema)
-                    sdf = self.spark.createDataFrame(data, schema=schema)
-                    self.assertEqual(cdf.schema, sdf.schema)
-                    self.assertEqual(cdf.collect(), sdf.collect())
-
-    def test_large_client_data(self):
-        # SPARK-42816 support more than 4MB message size.
-        # ~200bytes
-        cols = ["abcdefghijklmnoprstuvwxyz" for x in range(10)]
-        # 100k rows => 20MB
-        row_count = 100 * 1000
-        rows = [cols] * row_count
-        self.assertEqual(row_count, self.connect.createDataFrame(data=rows).count())
 
     def test_unsupported_jvm_attribute(self):
         # Unsupported jvm attributes for Spark session.
